math: move the shared RuleAnalysis contract out of the biological layer
RuleAnalysis is an algorithm over interfaces with one adapter per layer -
cbit.vcell.model.ModelRuleFactory over biological rules, cbit.vcell.math.MathRuleFactory
over ParticleJumpProcesses - and it references no concrete biological type. It was
simply declared in org.vcell.model.rbm, which forced the math adapter to import
from biology to implement it.

RuleAnalysis and RuleAnalysisReport move to cbit.vcell.math, the layer both
adapters can legally reach: MathRuleFactory now needs no import, ModelRuleFactory
is biological -> mathematical, NFsimXMLWriter is mathematical -> solver, and the
GUI callers are unaffected.

getNFSimXML (255 lines) does not move with it. It generates a solver input format,
so keeping it beside the contract would put solver-format generation in the math
layer - the P4 violation just fixed for the .lngv writers. It moves to
NFsimXMLWriter as a private static, where its only live caller already was.

Its other caller, RulebasedTransformer.compareOutputs, is deleted. It compared
BioNetGen's XML against ours and was unreachable - every call site commented out
in RulebasedMathMapping and LangevinMathMapping - and it was the only thing that
would have forced a new biological -> solver dependency. Git history keeps it.

Violations in cbit.vcell.math go from 9 to 1: only RbmUtils remains, whose paired
biological/math BNGL overloads are the other half of this work, now the sole entry
in the guard's accepted list.

Mechanical notes: getID was a same-class static call inside getNFSimXML, so 28
call sites needed qualifying after the split, and the nested contract types
(RuleEntry, ParticipantType, the bond entries, the Operation subtypes) each needed
explicit imports in RuleAnalysisReport and NFsimXMLWriter - nested types do not
follow the outer class.

vcell-core Fast group green; MathGen corpus 705 tests, 0 failures, which covers
the NFSim path this touches.
