+++ vcell-core/src/main/java/cbit/vcell/mapping/RulebasedTransformer.java
@@ -175,76 +175,6 @@ public class RulebasedTransformer implements SimContextTransformer {
 			return ruleElementMap;
 		}
 		
-		public void compareOutputs(MathRuleEntry mathRuleEntry) {
-			// compare the xml from BioNetGen with the one we build
-			XMLOutputter outp = new XMLOutputter();
-				
-			String identifier = RuleAnalysis.getID(mathRuleEntry);
-			Element rreTheirs = ruleElementMap.get(identifier);
-			
-			RuleAnalysisReport report = RuleAnalysis.analyze(mathRuleEntry, true);
-			Element rreOurs = RuleAnalysis.getNFSimXML(mathRuleEntry, report);
-
-			compareOperations(rreTheirs, rreOurs);
-			
-			rreTheirs.removeChild("RateLaw", Namespace.getNamespace("http://www.sbml.org/sbml/level3"));
-			rreTheirs.removeChild("ListOfOperations", Namespace.getNamespace("http://www.sbml.org/sbml/level3"));
-			rreOurs.removeChild("ListOfOperations");
-						
-			String rule_id_str = rreTheirs.getAttributeValue("id");
-			rule_id_str = rule_id_str.substring(2);
-			int rrIndex = Integer.parseInt(rule_id_str);
-			String sTheirs = outp.outputString(rreTheirs);
-			sTheirs = sTheirs.replace("xmlns=\"http://www.sbml.org/sbml/level3\"", "");
-			sTheirs = sTheirs.replaceAll("\\s+","");
-			
-//			String debugUser = PropertyLoader.getProperty("debug.user", "not_defined");
-//			if (debugUser.equals("danv") || debugUser.equals("mblinov"))
-//			{
-//				System.out.println("Saving their rules");
-//		        Format format = Format.getPrettyFormat();
-//				XMLOutputter outpEx = new XMLOutputter(format);
-//				String sTheirsEx = outpEx.outputString(rreTheirs);
-//				RulebasedTransformer.saveAsText("c:\\TEMP\\ddd\\theirRules.txt", sTheirsEx);
-//			}
-
-			String sOurs = outp.outputString(rreOurs);
-			sOurs = sOurs.replace(".0", "");
-			sOurs = sOurs.replaceAll("\\s+","");
-			
-//			if (debugUser.equals("danv") || debugUser.equals("mblinov"))
-//			{
-//				System.out.println("Saving our rules");
-//		        Format format = Format.getPrettyFormat();
-//				XMLOutputter outpEx = new XMLOutputter(format);
-//				String sOursEx = outpEx.outputString(rreOurs);
-//				RulebasedTransformer.saveAsText("c:\\TEMP\\ddd\\ourRules.txt", sOursEx);
-//			}
-			
-			if(!sTheirs.equals(sOurs)) {
-//				System.out.println(sTheirs);
-//				System.out.println(sOurs);
-				char[] sTheirsChars = sTheirs.toCharArray();
-				char[] sOursChars = sOurs.toCharArray();
-				int firstMismatchIndex = -1;
-				int minLength = Math.min(sOurs.length(),sTheirs.length());
-				for (int i=0;i<minLength;i++){
-					if (sTheirsChars[i] != sOursChars[i]){
-						firstMismatchIndex = i;
-						break;
-					}
-				}
-				
-				int startingIndex = Math.max(0, firstMismatchIndex-500);
-				int endingIndexOurs = Math.min(sOurs.length(), firstMismatchIndex+500);
-				int endingIndexTheirs = Math.min(sTheirs.length(), firstMismatchIndex+500);
-//				System.out.println("not matching!!!");
-				throw new RuntimeException("Rule not matching\n" + sTheirs.substring(startingIndex, endingIndexTheirs) + "\n" + sOurs.substring(startingIndex, endingIndexOurs));
-//				BeanUtils.sendRemoteLogMessage(null, sTheirs + "\n" + sOurs);
-			} else {
-//				System.out.println("good match ");
-			}
-		}
 	}
 
 	// we populate these by parsing the xml file produced by BioNetGen
