+++ vcell-core/src/main/java/cbit/vcell/math/RuleAnalysis.java
@@ -1,4 +1,4 @@
-package org.vcell.model.rbm;
+package cbit.vcell.math;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -11,13 +11,6 @@ import java.util.TreeSet;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jdom2.Element;
-import org.vcell.model.rbm.RuleAnalysisReport.AddBondOperation;
-import org.vcell.model.rbm.RuleAnalysisReport.AddMolecularTypeOperation;
-import org.vcell.model.rbm.RuleAnalysisReport.ChangeStateOperation;
-import org.vcell.model.rbm.RuleAnalysisReport.DeleteBondOperation;
-import org.vcell.model.rbm.RuleAnalysisReport.DeleteMolecularTypeOperation;
-import org.vcell.model.rbm.RuleAnalysisReport.DeleteParticipantOperation;
-import org.vcell.model.rbm.RuleAnalysisReport.Operation;
 
 public class RuleAnalysis {
 
@@ -570,261 +563,4 @@ public class RuleAnalysis {
 		int componentIndex = mce.getComponentIndex();
 		return getID(mce.getMolecularTypeEntry())+"_"+"C"+(componentIndex+INDEX_OFFSET);
 	}
-	
-	public static Element getNFSimXML(RuleEntry rule, RuleAnalysisReport report){
-		Element root = new Element("ReactionRule");
-		root.setAttribute("id",getID(rule));
-		root.setAttribute("name",rule.getRuleName());
-		root.setAttribute("symmetry_factor",""+report.getSymmetryFactor());
-		
-		Element listOfReactantPatterns = new Element("ListOfReactantPatterns");
-		root.addContent(listOfReactantPatterns);
-		for (ParticipantEntry reactantEntry : rule.getReactantEntries()){
-			Element reactantElement = getParticipantEntry(rule,reactantEntry);
-			listOfReactantPatterns.addContent(reactantElement);
-		}
-		if(rule.getReactantEntries().isEmpty()) {
-			listOfReactantPatterns.addContent(" ");	// we want <ListOfReactantPatterns> </ListOfReactantPatterns> rather than <ListOfReactantPatterns/>
-		}
-		
-		Element listOfProductPatterns = new Element("ListOfProductPatterns");
-		root.addContent(listOfProductPatterns);
-		for (ParticipantEntry productEntry : rule.getProductEntries()){
-			Element productElement = getParticipantEntry(rule,productEntry);
-			listOfProductPatterns.addContent(productElement);
-		}
-		if(rule.getProductEntries().isEmpty()) {
-			listOfProductPatterns.addContent(" ");	// we want <ListOfProductPatterns> </ListOfProductPatterns> rather than <ListOfProductPatterns/>
-		}
-
-		
-//		Element rateLaw = new Element("RateLaw");
-//		root.addContent(rateLaw);
-//		Element listOfRateConstants = new Element("ListOfRateConstants");
-//		rateLaw.addContent(listOfRateConstants);
-//		Element rateConstant = new Element("RateConstant");
-//		listOfRateConstants.addContent(rateConstant);
-//		rateConstant.setAttribute("value",rule.getForwardRateConstantName());
-		
-		// we build a map of strings with the source ID as keys and target ID as content
-		Map<String, String> idMap = new LinkedHashMap<String, String>();
-		for (MolecularTypeEntry ractantMolecule : report.getMappedReactantMolecules()){
-			MolecularTypeEntry productMolecule = report.getMappedProductMolecules(ractantMolecule).get(0);
-			if(productMolecule == null) {
-				idMap.put(getID(ractantMolecule), null);
-			} else {
-				idMap.put(getID(ractantMolecule), getID(productMolecule));
-			}
-			for (MolecularComponentEntry reactantComponent : ractantMolecule.getMolecularComponentEntries()){
-				MolecularComponentEntry productComponent = report.getMappedProductComponent(reactantComponent);
-				if(productComponent == null) {
-					idMap.put(getID(reactantComponent), null);
-				} else {
-					idMap.put(getID(reactantComponent), getID(productComponent));
-				}
-			}
-		}
-		Element map = new Element("Map");
-		root.addContent(map);
-		SortedSet<String> keys = new TreeSet<String>(idMap.keySet());	// we create the map sorted by sourceID
-		for (String sourceID : keys) {
-			String targetID = idMap.get(sourceID);
-			Element mapItem = new Element("MapItem");
-			map.addContent(mapItem);
-			mapItem.setAttribute("sourceID", sourceID);
-			if(targetID != null) {
-				mapItem.setAttribute("targetID", targetID);
-			}
-		}
-		if(keys.isEmpty()) {
-			map.addContent(" ");	// we want <Map> </Map> rather than <Map/>
-		}
-	
-		
-//		Element map = new Element("Map");
-//		root.addContent(map);
-//		for (MolecularTypeEntry mappedReactantMolecule : report.getMappedReactantMolecules()){
-//			Element moleculeMapEntry = new Element("MapItem");
-//			map.addContent(moleculeMapEntry);
-//			MolecularTypeEntry productMolecule = report.getMappedProductMolecules(mappedReactantMolecule).get(0);
-//			moleculeMapEntry.setAttribute("sourceID",getID(mappedReactantMolecule));
-//			if(productMolecule != null) {
-//				moleculeMapEntry.setAttribute("targetID",getID(productMolecule));
-//			}
-//			for (MolecularComponentEntry reactantComponent : mappedReactantMolecule.getMolecularComponentEntries()){
-//				Element componentMapEntry = new Element("MapItem");
-//				map.addContent(componentMapEntry);
-//				MolecularComponentEntry productComponent = report.getMappedProductComponent(reactantComponent);
-//				componentMapEntry.setAttribute("sourceID",getID(reactantComponent));
-//				if(productComponent != null) {
-//					componentMapEntry.setAttribute("targetID",getID(productComponent));
-//				}
-//			}
-//		}
-		
-		Element listOfOperations = new Element("ListOfOperations");
-		root.addContent(listOfOperations);
-		if(report.getOperations().isEmpty()) {
-			listOfOperations.addContent(" ");	// we want <ListOfOperations> </ListOfOperations> rather than <ListOfOperations/>
-		}
-		
-		for (Operation op : report.getOperations()){
-			if (op instanceof ChangeStateOperation){
-				ChangeStateOperation changeStateOp = (ChangeStateOperation)op;
-				Element changeState = new Element("StateChange");
-				listOfOperations.addContent(changeState);
-				changeState.setAttribute("site",getID(changeStateOp.reactantComponentEntry));
-				changeState.setAttribute("finalState",changeStateOp.newState);
-			}
-		}
-		for (Operation op : report.getOperations()){
-			if (op instanceof DeleteBondOperation){
-				DeleteBondOperation deleteBondOp = (DeleteBondOperation)op;
-				Element deleteBond = new Element("DeleteBond");
-				listOfOperations.addContent(deleteBond);
-				deleteBond.setAttribute("site1",getID(deleteBondOp.removedBondEntry.reactantComponent1));
-				deleteBond.setAttribute("site2",getID(deleteBondOp.removedBondEntry.reactantComponent2));
-			}
-		}
-		for (Operation op : report.getOperations()){
-			if (op instanceof AddBondOperation){
-				AddBondOperation addBondOp = (AddBondOperation)op;
-				Element addBond = new Element("AddBond");
-				listOfOperations.addContent(addBond);
-				addBond.setAttribute("site1",getID(addBondOp.addedProductBondEntry.reactantComponent1));
-				addBond.setAttribute("site2",getID(addBondOp.addedProductBondEntry.reactantComponent2));
-			}
-		}
-		for (Operation op : report.getOperations()){
-			if (op instanceof AddMolecularTypeOperation){
-				AddMolecularTypeOperation addMoleculeOp = (AddMolecularTypeOperation)op;
-				Element addMolecule = new Element("Add");
-				listOfOperations.addContent(addMolecule);
-				addMolecule.setAttribute("id",getID(addMoleculeOp.unmatchedProductMoleculeEntry));
-			}
-		}
-		for (Operation op : report.getOperations()){
-			if (op instanceof DeleteMolecularTypeOperation){
-				DeleteMolecularTypeOperation deleteMolecule = (DeleteMolecularTypeOperation)op;
-				Element delete = new Element("Delete");
-				listOfOperations.addContent(delete);
-				delete.setAttribute("id",getID(deleteMolecule.removedReactantMolecularEntry));
-				delete.setAttribute("DeleteMolecules","1");
-			}
-		}
-		for (Operation op : report.getOperations()){
-			if (op instanceof DeleteParticipantOperation){
-				DeleteParticipantOperation deleteParticipant = (DeleteParticipantOperation)op;
-				Element delete = new Element("Delete");
-				listOfOperations.addContent(delete);
-				delete.setAttribute("id",getID(deleteParticipant.removedParticipantEntry));
-				delete.setAttribute("DeleteMolecules","0");
-			}
-		}
-		return root;
-	}
-
-	private static Element getParticipantEntry(RuleEntry rule, ParticipantEntry participant) {
-		Element root = null;
-		if (participant.getParticipantType() == ParticipantType.Reactant){
-			root = new Element("ReactantPattern");
-		}else{
-			root = new Element("ProductPattern");
-		}
-		root.setAttribute("id", getID(participant));
-		Element listOfMolecules = new Element("ListOfMolecules");
-		root.addContent(listOfMolecules);
-		for (MolecularTypeEntry molecule : participant.getMolecularTypeEntries()){
-			Element moleculeElement = new Element("Molecule");
-			listOfMolecules.addContent(moleculeElement);
-			moleculeElement.setAttribute("id",getID(molecule));
-			moleculeElement.setAttribute("name",molecule.getMolecularTypeName());
-			if (molecule.getMatchLabel()!=null){
-				moleculeElement.setAttribute("label",molecule.getMatchLabel());
-			}
-			if (molecule.getMolecularComponentEntries().size()>0){
-				Element listOfComponents = new Element("ListOfComponents");
-				moleculeElement.addContent(listOfComponents);
-				for (MolecularComponentEntry component : molecule.getMolecularComponentEntries()){
-					boolean bStateSpecified = false;
-					boolean bBondSpecified = false;
-					Element componentElement = new Element("Component");
-					componentElement.setAttribute("id",getID(component));
-					componentElement.setAttribute("name",component.getMolecularComponentName());
-					String state = component.getExplicitState();
-					if (state != null){
-						bStateSpecified = true;
-						componentElement.setAttribute("state",state);
-					}
-					if(component.isBondExists()) {
-						componentElement.setAttribute("numberOfBonds","+");
-						bBondSpecified = true;
-					} else if (component.isBondPossible()){
-						componentElement.setAttribute("numberOfBonds","?");
-					} else if (component.hasBond()){
-						componentElement.setAttribute("numberOfBonds","1");
-						bBondSpecified = true;
-					} else {
-						componentElement.setAttribute("numberOfBonds","0");
-						bBondSpecified = true;
-					}
-					if (bStateSpecified || bBondSpecified){
-						listOfComponents.addContent(componentElement);
-					}
-				}
-			}
-		}
-		Element listOfBonds = new Element("ListOfBonds");
-		boolean bAnyBonds = false;
-		if (participant.getParticipantType() == ParticipantType.Reactant){
-			int count = 0;
-			Map<String,String> bondSiteMap = new HashMap<String, String>();
-			for (ReactantBondEntry bond : rule.getReactantBondEntries()){
-				if (bond.reactantComponent1.getMolecularTypeEntry().getParticipantEntry() == participant){
-					if(bond.reactantComponent1 == null || bond.reactantComponent2 == null) {
-						throw new RuntimeException("Bond site(s) are null when trying to build the bond map.");
-					}
-					bondSiteMap.put(getID(bond.reactantComponent1), getID(bond.reactantComponent2));
-				}
-			}
-			SortedSet<String> keys = new TreeSet<String>(bondSiteMap.keySet());
-			for (String key : keys) { 
-				String value = bondSiteMap.get(key);
-				Element bondElement = new Element("Bond");
-				listOfBonds.addContent(bondElement);
-				bondElement.setAttribute("id",getID(participant)+"_B"+(count+INDEX_OFFSET));
-				bondElement.setAttribute("site1", key);
-				bondElement.setAttribute("site2", value);
-				bAnyBonds = true;
-				count++;
-			}
-		}else{
-			int count = 0;
-			Map<String,String> bondSiteMap = new HashMap<String, String>();
-			for (ProductBondEntry bond : rule.getProductBondEntries()){
-				if (bond.productComponent1.getMolecularTypeEntry().getParticipantEntry() == participant){
-					if(bond.productComponent1 == null || bond.productComponent2 == null) {
-						throw new RuntimeException("Bond site(s) are null when trying to build the bond map.");
-					}
-					bondSiteMap.put(getID(bond.productComponent1), getID(bond.productComponent2));
-				}
-			}
-			SortedSet<String> keys = new TreeSet<String>(bondSiteMap.keySet());
-			for (String key : keys) { 
-				String value = bondSiteMap.get(key);
-				Element bondElement = new Element("Bond");
-				listOfBonds.addContent(bondElement);
-				bondElement.setAttribute("id",getID(participant)+"_B"+(count+INDEX_OFFSET));
-				bondElement.setAttribute("site1", key);
-				bondElement.setAttribute("site2", value);
-				bAnyBonds = true;
-				count++;
-			}
-		}
-		if (bAnyBonds){
-			root.addContent(listOfBonds);
-		}
-		return root;
-	}
-	
 }
